block admin from deleting a laptop that is currently present in a customers basket

+++ delete.php
@@ -4,12 +4,19 @@ require_once('connect.php');
 $id = $_GET['id'];
 
 try {
-   $stmt = $conn->prepare('DELETE FROM laptops WHERE id=:id');
+   // check if laptop is in basket table, if so prevent deleting it.
+   $stmt = $conn->prepare('SELECT * FROM basket WHERE product_id=:id');
    $stmt->execute(['id' => $id]);
+   if ($stmt->rowCount()>0) {
+      echo "<script>alert('Product found in basket. Deletion not allowed!');</script>";
+   } else {
+      $stmt = $conn->prepare('DELETE FROM laptops WHERE id=:id');
+      $stmt->execute(['id' => $id]);   
+   }
    }
 catch(PDOException $e)
    {
-   echo 'DELETE FROM query' . "<br>" . $e->getMessage();
+   echo $e->getMessage();
    }
 $conn = null;
